first pass at cleaning up extraction of `Interactable` and `listener`
