+++ src/interact.js
@@ -554,72 +554,6 @@
                     : scope.matchesUpTo(element, value, interactableElement);
     }
 
-    function defaultActionChecker (pointer, interaction, element) {
-        var rect = this.getRect(element),
-            shouldResize = false,
-            action = null,
-            resizeAxes = null,
-            resizeEdges,
-            page = utils.extend({}, interaction.curCoords.page),
-            options = this.options;
-
-        if (!rect) { return null; }
-
-        if (scope.actionIsEnabled.resize && options.resize.enabled) {
-            var resizeOptions = options.resize;
-
-            resizeEdges = {
-                left: false, right: false, top: false, bottom: false
-            };
-
-            // if using resize.edges
-            if (scope.isObject(resizeOptions.edges)) {
-                for (var edge in resizeEdges) {
-                    resizeEdges[edge] = checkResizeEdge(edge,
-                                                        resizeOptions.edges[edge],
-                                                        page,
-                                                        interaction._eventTarget,
-                                                        element,
-                                                        rect,
-                                                        resizeOptions.margin || scope.margin);
-                }
-
-                resizeEdges.left = resizeEdges.left && !resizeEdges.right;
-                resizeEdges.top  = resizeEdges.top  && !resizeEdges.bottom;
-
-                shouldResize = resizeEdges.left || resizeEdges.right || resizeEdges.top || resizeEdges.bottom;
-            }
-            else {
-                var right  = options.resize.axis !== 'y' && page.x > (rect.right  - scope.margin),
-                    bottom = options.resize.axis !== 'x' && page.y > (rect.bottom - scope.margin);
-
-                shouldResize = right || bottom;
-                resizeAxes = (right? 'x' : '') + (bottom? 'y' : '');
-            }
-        }
-
-        action = shouldResize
-            ? 'resize'
-            : scope.actionIsEnabled.drag && options.drag.enabled
-                ? 'drag'
-                : null;
-
-        if (scope.actionIsEnabled.gesture
-            && interaction.pointerIds.length >=2
-            && !(interaction.dragging || interaction.resizing)) {
-            action = 'gesture';
-        }
-
-        if (action) {
-            return {
-                name: action,
-                axis: resizeAxes,
-                edges: resizeEdges
-            };
-        }
-
-        return null;
-    }
 
     var InteractEvent = require('./InteractEvent');
 
@@ -909,7 +843,7 @@
             interactables         : scope.interactables,
             pointerIsDown         : interaction.pointerIsDown,
             defaultOptions        : scope.defaultOptions,
-            defaultActionChecker  : defaultActionChecker,
+            defaultActionChecker  : require('./defaultActionChecker'),
 
             actionCursors         : scope.actionCursors,
             dragMove              : listener.listeners.dragMove,
@@ -1083,12 +1017,6 @@
             };
         };
     };
-
-    function endAllInteractions (event) {
-        for (var i = 0; i < scope.interactions.length; i++) {
-            scope.interactions[i].pointerEnd(event, event);
-        }
-    }
 
     listener.listenToDocument(scope.document);
